requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
Django==3.0.6
mariadb==0.9.58
mysqlclient==1.4.6
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
